[MIRROR] [MDB IGNORE] allows janitor keys to be stored in janitor wintercoats and janibets (#8055)
Original PR: https://github.com/tgstation/tgstation/pull/78724
-----
## About The Pull Request
the janitor access key could not be stored on their belt, or their
wintercoat slot, which felt off to me. so i added it to both of them,
and added the janicart key to the coat slot as well. i figure if one can
fit, the other probably can too.
## Why It's Good For The Game
convenient, less cumbersome storage for items that you want to keep both
close to your person, and accessible for easy use.
## Changelog
:cl:
qol: allows janitor keys to be stored in janitor wintercoats and
janibets
/:cl:

---------

Co-authored-by: Lamb <110322848+CoiledLamb@users.noreply.github.com>