[MIRROR] [MDB IGNORE] Recaches the icons generated in the `changeling` unit test (#8069)
Original PR: https://github.com/tgstation/tgstation/pull/78799
-----
## About The Pull Request
Title.

## Why It's Good For The Game
Fixes: #78673 

Hi. I run one of the two downstreams and was working on Melberts ling
transformation PR.1

Turns out, if you have a lot of DNA features that have a lot of icons
associated with them, this unit test just shits itself with a bad icon
operation.

I spoke with Lemons over this and he helped point me in the right
direction, which lead to recaching the icons to prevent issues.

As it stands, this *probably* wouldn't happen on base /tg/ since DNA
features and their associated parts aren't a lot (compared to
elsewhere). But if/when more features gets added this should prevent any
issues.

## Changelog
Not player facing.

---------

Co-authored-by: Jolly <70232195+Jolly-66@users.noreply.github.com>