[MIRROR] [MDB IGNORE] Adds a smidge of information for species hard-dels (#8251)
Original PR: https://github.com/tgstation/tgstation/pull/79064
-----
## About The Pull Request

Species hard deletes are a little difficult to track down without the
ref tracker due to having no owner so I figure this might help... just a
tiny bit. Maybe.

This will let us know if the species hard-deleting was applied to a
human or not. `properly_gained` is set to `TRUE` in `on_species_gain`.
If it is false, it might be a shadow species kept in DNA or some other
weird place or something.

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>
Co-authored-by: Jolly <70232195+Jolly-66@users.noreply.github.com>