[MIRROR] [MDB IGNORE] resprites air alarms (#8316)
Original PR: https://github.com/tgstation/tgstation/pull/79016
-----
## About The Pull Request
foreword: this will not negatively impact the wallening
okay now that we've got that out of the way - 

![image](https://github.com/tgstation/tgstation/assets/110322848/c3dd36a5-5ed0-46d3-94b3-d4504a0a7409)
resprites air alarms to be closer to the older versions, and the
wallening vers.

![image](https://github.com/tgstation/tgstation/assets/110322848/c7bec0df-1110-4af3-8b8c-4e73c2c3eddc)
## Why It's Good For The Game
the current air alarms are _fine_, but they carry with them some bad
habits - grey outlines on the body, and unnecessary animations on the
screens. as well as being obfuscated by having the safe reading being
blue, rather than green. this will help to make things more contiguous
going in to the wallening, but will also serve for some quality of life
benefits, as the readout will be much clearer.
## Changelog
:cl:
image: resprites air alarms
/:cl:

Co-authored-by: Lamb <110322848+CoiledLamb@users.noreply.github.com>