[MIRROR] [MDB IGNORE] Minor signals memory optimization (#8319)
Original PR: https://github.com/tgstation/tgstation/pull/79153
-----
This replaces the inner list of signals which tracked receivers with a
flat list instead of the previous keyed list that was used to prevent
duplicates. We have code already checking for duplicates before hand so
we can lean on that to assume the sender already has the receiver
listed. This should also be minutely more performant equal to about a
single if statement but I don't want to bother trying to profile
something so small.

---------

Co-authored-by: Emmett Gaines <ninjanomnom@gmail.com>