[MIRROR] [MDB IGNORE] Adds Examine hints for Railing Deconstruction (#8325)
Original PR: https://github.com/tgstation/tgstation/pull/79205
-----

## About The Pull Request
While railings already have contextual tips, the context requires you to
already know what tool to be holding. Would probably be helpful to know
at a glance which tool that is.


![image](https://github.com/tgstation/tgstation/assets/76465278/7626b3d9-2c18-4c82-b6e7-44423027914c)
## Why It's Good For The Game
Usability? Is that the word? IDK I noticed it was missing as I kept
hitting it with a screwdriver so I'm sure I can't be the only one.
## Changelog
:cl:
qol: Railings now have Examine hints for how to deconstruct them
/:cl:

---------

Co-authored-by: OrionTheFox <76465278+OrionTheFox@users.noreply.github.com>