[MIRROR] [MDB IGNORE] Modular Shield Generators now play well with shuttles [NO GBP] (#8406)
Original PR: https://github.com/tgstation/tgstation/pull/79256
-----

## About The Pull Request
So an issue with the shield generator and shuttles is that on some maps
depending on how the shuttle docks, a change of direction is possible on
launch and my dumb brain made it so that the machines lose their links
on change direction so I moved that to the wrench act.

I also made it so the shield generator automatically turns off when it
changes z levels due to shuttle movement so it doesnt leave forcefield
tiles behind, it can be turned back on immediately after. I also updated
the description of the generator so this isnt a hidden function.
## Why It's Good For The Game
Less buggy system and now plays well with things like white ship and
evac shuttle. Modular shield generators are not designed as something
you are meant to have to or even be able to individually reach all of
their parts to relink/fix unless a freak accident or sabotage happened
and existing on a shuttle is not a freak accident or sabotage.
## Changelog
:cl:
fix: Modular shield generator modules no longer lose linkage when riding
a shuttle
fix: Modular shield generators now gracefully turn off when being moved
by a shuttle rather than leaving their projections behind, the
generator's description was updated to advertise this behavior.
/:cl:

---------

Co-authored-by: moocowswag <62126254+moocowswag@users.noreply.github.com>