[MIRROR] [MDB IGNORE] Fixes mafia achievements desyncing you from the database (#8486)
Original PR: https://github.com/tgstation/tgstation/pull/79555
-----
I swear to fucking god

PDA mafia has two bodies, the real player body, and the fake mafia
stand-in. When an achievement is obtained, it's marked as achieved
directly on the players client, but then passes the empty mafia stand-in
for any afterwork. This causes achievement notifications to be muted,
the code to runtime and for it not to update every appropriatew database
table (like the achievement highscore table).

:cl:
Fix: Fixes PDA mafia achievements desyncing you from the database
/:cl:

On a somewhat related note, anyone that has gotten PDA mafia
achievements (which includes me) is now desynced from the
ACHIEVEMENTS_SCORE table. Looks like the only fix is to reset the score
in the ACHIEVEMENTS_SCORE to a non-numerical value and force the game to
recount every achievement

Ghommie you made the scoreboard, is the best solution here to just wipe
"Achievements Score" and force the game to recount them?

---------

Co-authored-by: Time-Green <7501474+Time-Green@users.noreply.github.com>