Fixes Wiki Book Manuals not working (#79589)
## About The Pull Request

Fixes #79545

Basically, the TGUI markdown reader that every other book uses didn't
properly send the HTML printified wiki page that we had. In order to fix
this, I shuffled around a bunch of code and worked in an edge case to
ensure that these manuals keep functioning.

In an ideal world we would be able to do something more elegant, or
maybe not even rely on displaying wiki pages in-game at all, but we just
gotta play the hand we're dealt and ensure that these sorts of features
continue to work for right now.

Also, we no longer need to both write and access raw HTML in the book
framework to make everything work here. This should reduce a LOT of
potential exploit vectors.

## Why It's Good For The Game


![image](https://github.com/tgstation/tgstation/assets/34697715/061ed8d1-2f33-4482-977e-e7db5ad59672)

Thing work 👍 
## Changelog
:cl:
fix: Nanotrasen has clarified an issue with their manual publishers, and
these guides should now contain actual user-pertinent content.
/:cl: