Fixes an oversight in the examine message for a carbon with an empty golem stomach. (#79657)
## About The Pull Request

Noticed on a round recently where the warden got the golem stomach that
the message read as "He are as still as a statue!"
Turns out that the "are" wasn't setup in our pronoun system properly. I
think this wasn't noticed before because golems don't have a gender,
maybe?
## Why It's Good For The Game

Grammatical errors? bad!
## Changelog
:cl:DATA_
spellcheck: The examine message for a carbon with an empty golem stomach
now properly matches said carbon's gender.
/:cl: