Fixes Ashwalker Tendril Revives (#79637)
## About The Pull Request
- Changes the Ashwalker Tendril revive proc to hold the old body until
the revive is done, and only delete the old body after moving the
mind/client over.
- Changes a single-letter var name to be more proper.
## Why It's Good For The Game
Fixes #79635

The issue seems to have been due to the old mob being deleted, which
deleted the mind. Later, when the new body was created, the mind was
unavailable due to being deleted. So now, we just hold the old body for
a bit and delete it once the new body is finished.

I'm not sure why this suddenly stopped working. Perhaps we were abusing
a long hard-del that delayed the mind being deleted and now it's no
longer a thing?
## Changelog
:cl:
fix: Ashwalker Tendril Revives no longer ghost the player being revived.
/:cl: