[MIRROR] [MDB IGNORE] Fixes the EMP Barsign sprite not displaying (#8569)
Original PR: https://github.com/tgstation/tgstation/pull/79599
-----
## About The Pull Request

The logic that sets the icon_state had an error in it. When the sign got
emp'd, they'd get the `NOPOWER` flag which was causing them to get the
`empty` `icon_state` instead of that of the emped barsign.

This just fixes that by adding a conditional check for `EMPED`, and also
resprites the emissive lightmask for the emped barsign to match the
individual signs that make up its sprite, where previously it was just a
static square.

Also added it to the barsigns unit tests because I can see someone
breaking this again.

## Why It's Good For The Game

Fixes bug where emped barsigns would show the turned off sprite instead
of their animated sprite. Also improves the light mask:

Note: this is in complete darkness. The colored backdrops of the signs
are not supposed to be emissive.

<details> <summary>Before </summary>


![dreamseeker_jiXgQtteGK](https://github.com/tgstation/tgstation/assets/13398309/4c576c3d-8758-4ce8-b1b5-00d0fb341f38)

</details>

<details> <summary>After </summary>


![dreamseeker_tCu8YSTJuZ](https://github.com/tgstation/tgstation/assets/13398309/10c328c9-8e99-47c7-98a4-e6f66c4e5eee)

</details>


![dreamseeker_Lu3TfIzkWO](https://github.com/tgstation/tgstation/assets/13398309/c7251cc0-84ab-432d-85e9-2b93b94b63e4)

## Changelog

:cl:
fix: emped bar signs will now display the correct sprite
image: added a more detailed lightmask for the emp bar sign sprite
/:cl:

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>