[MIRROR] [MDB IGNORE] Fixes a small issue with disposal outlets. (#8623)
Original PR: https://github.com/tgstation/tgstation/pull/79706
-----
## About The Pull Request
Disposal outlets eject the contents of the holder with a 2 seconds
delay. Delete the outlet before the delay is over and the contents will
stay stuck inside the holder.

Also merging disposal holders doesn't properly register a signal for
mobs transfered to the new holder.

## Why It's Good For The Game
Bugfixing. Disposal code is a mess. Perhaps related to #79629 but I
couldn't exactly reproduce it.

## Changelog

:cl:
fix: Fixed a small issue with disposal outlets leaving contents about to
be ejected stuck inside the pipe beneath it if deleted.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>