[MIRROR] [MDB IGNORE] General code maintenance for smart fridge (#8625)
Original PR: https://github.com/tgstation/tgstation/pull/79623
-----
## About The Pull Request
- Fixes #77805
- Fixes #77593

Overall code optimizations for smart fridges. The gist of it includes.
1. Converted UI to typescript
2. Smart fridges won't be accepting items that are either abstract or
holograms
3. Autodoc for procs & vars
4. Merged a lot of if statements into 1 ternary operation. Used
`is_type_in_list()` in some places to reduce the redundant `istype()`
checks
5. Removed all single letter var names
6. Moved all subtype specific vars and functions away from the parent
fridge type. If we need a subtype to perform a certain operation the
parent type should provide vars & procs to override and we should not be
checking for child types inside the parent type. In this case it was
overlays for the contents displayed on the fridge & examines

Qol improvements include.
1. Added better examine & screen tips for fridge's
2. Drying rack can now be deconstructed with a crowbar and not simply
pried open with it.

## Changelog
:cl:
fix: drying rack now shows correct examines & screen tips.
code: tone of code organization for smart fridges overall. changed ui to
typescript.
qol: added more detailed examines & screen tips for smart fridges.
drying racks can be dismantled with a crowbar and not simply pried open
with it.
/:cl:

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>