[MIRROR] [MDB IGNORE] Adds a toggle action to implant HUDs (#8673)
Original PR: https://github.com/tgstation/tgstation/pull/79777
-----

## About The Pull Request
Adds an action for all implanted HUDs to toggle them on and off. The
default state is on, and the state is toggled appropriately when the
implant is added or removed. This is done by using hud.hide_from and
hud.show_to.
## Why It's Good For The Game
Quality of life change for roleplay situations where the HUD is usless
clutter.
## Changelog
:cl:
qol: Implanted HUDs can now be toggled on and off with an action.
/:cl:

---------

Co-authored-by: Shroopy <delyth@shroop.net>
Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>