[MIRROR] [MDB IGNORE] Enzyme (var) exorcism in Metastation (#8712)
Original PR: https://github.com/tgstation/tgstation/pull/79807
-----

## About The Pull Request

Do not ask why, do not ask how, but the Universal Enzyme in
metastation's kitchen was varedited to be on a higher layer. I exorcised
that var out of the bottle and it should be a normal bottle now.

## Why It's Good For The Game
That thing should not be varedited like that. Also moved it slightly to
look slightly better.

## Changelog
:cl:
qol: Slightly moved the universal enzyme on meta's kitchen to a prettier
spot.
fix: The universal enzyme on meta's kitchen is no longer unecessarely
varedited.
/:cl:

---------

Co-authored-by: Fazzie <84548101+EuSouAFazer@users.noreply.github.com>