[MIRROR] [MDB IGNORE] Makes (Allergy) Dogtags Tiny items, able to fit into wallets. (#8719)
Original PR: https://github.com/tgstation/tgstation/pull/79843
-----
## About The Pull Request
Title.

~~For future thought: maybe examining (double-examining?) somebody with
a tag in their wallet (in the ID slot) or attached to their uniform
should show what it says? Would make it easier to read.~~
Okay so it DOES work on uniforms. But still, if somebody better at
coding wants to try to make it also do that in wallet that'd be neat.
## Why It's Good For The Game
I mean, it's not like it hurts. 
Plus, you can fit a lighter, but not a little tag?
## Changelog
:cl:
qol: Allergy Dogtags (and any other dogtags, really) are now Tiny items
and can fit into wallets.
/:cl:

---------

Co-authored-by: OrionTheFox <76465278+OrionTheFox@users.noreply.github.com>