Makes gas visuals send less assets (#78955)
![Discord_2023-10-12_13-21-06](https://github.com/tgstation/tgstation/assets/1234602/e13cb015-dd94-48d9-a593-4258609049d1)

~~I'll see how this works out for the downstream, if it improves the
situation for them I'll check the profiler to see how much it affects
performance.~~

This reduces assets that need to be sent to the client, how much is hard
to tell but a downstream that was having issues due to it got reduced
issues when testing with this. Server side costs are roughly equivalent
now that I switched to using a straight color in the color matrix
filter.