Fixes Ranged Guardians from Shooting while Incorporeal (#79925)
## About The Pull Request

Fixes #79921

Otherwise, on the tin. Attack mode is for attacking, scouting mode is
for scouting. We were listening for clicking and stuff like that but it
was still failing somehow so this is le fix
## Why It's Good For The Game

You aren't supposed to shoot in this mode, only scout. 
## Changelog
:cl:
fix: Ranged Guardians (Holoparasites/Power Miners/etc.) can no longer
use ranged attacks in scouting (incorporeal) mode.
/:cl: