[MIRROR] [MDB IGNORE] Add a bounds check to prevent runtimes on the ripple shuttle port (#8810)
Original PR: https://github.com/tgstation/tgstation/pull/79959
-----
This appears to sometimes runtime when the target stationary port is
somehow a different size to the ripple port. The code had an out of
bounds check of a sort, but it doesn't actually work because the array
accesses runtime instead of returning a null

As to why the turfs available were not the same? most likely because the
bounds overlapped the world edge due to a fuck huge shuttle.

I have a feeling even if this proc completes something would fail later
anyway.

edit: oh so it was the lance shuttle which is fuck hueg

#79961 is related but not fixed by this.

---------

Co-authored-by: oranges <email@oranges.net.nz>