[MIRROR] [MDB IGNORE] Makes the QM Skirt Work For Digi Legs (#8879)
Original PR: https://github.com/tgstation/tgstation/pull/80067
-----

## About The Pull Request
Makes digi legs use the QM skirt properly.
<img src="https://i.ibb.co/9qKDkr3/Digi-Lizard-Skirt.png">
(Difference illustrated with VV)
## Why It's Good For The Game
Realistically lizards shouldn't be allowed in this role, as such
inferior beings have no business being a head of staff on a cutting-edge
Nanotrasen research station. But if a human decides to get a leg
transplant or something this is good for them.
Clearly an oversight in the code. Fixes good oversights bad for the
project.
## Changelog
:cl:
fix: Digi legs work with the QM's jumpskirt
/:cl:

---------

Co-authored-by: Jackraxxus <60418544+Jackraxxus@users.noreply.github.com>