Clown Car Steals Accesses Again (#80288)
## About The Pull Request

So, some time ago we fiddled around with vehicles to make them work
better, and we accidentally lost what was supposed to be an intentional
gimmick of the clown car. Clown cars used to obtain the accesses of
victims that got kidnapped into the car, but this hasn't worked for a
while now (supposedly around 15 months). This PR re-adds that behavior.

## Why It's Good For The Game

Access snatching is a fun gimmick, and it should strike fear into
everyone's hearts when the clown car finally catches the captain or a
station engineer.

## Changelog

:cl:
fix: The clown car once again obtains the accesses of kidnapped crew
within the car. Supermatter crystals beware!
/:cl: