General maintenance for chem heater (#80495)
## About The Pull Request

1. Converted UI to typescript
2. Removed all tutorial code(and the icon files it used) Also removes
the reward that came with it. Both were not a big deal to begin with so
you know no one will miss it, In the name of code clean up and reducing
dmi file sizes it serves us better removing them

    It did not compile in the new UI
![Screenshot
(368)](https://github.com/tgstation/tgstation/assets/110812394/cff143c8-5239-45e0-932d-df6447ab359b)
And the text to display is way too clunky to be readable code wise (no
indentation just big raw blobs of unorganized text)

https://github.com/tgstation/tgstation/blob/9983b883f7c0b52697f0c9f3fe5b483cba25d4ff/code/modules/reagents/chemistry/machinery/chem_heater.dm#L334-L339
    
And the steps manually set chem temps and volumes without calling the
appropriate procs basically its a mess so best to throw it all away.

3. Fixed the debug chem heater only absorbing 100 units of buffer from
the beaker even though it has a capacity of 2000. This is because it
used a constant of 100 for calculating how much volume to absorb

https://github.com/tgstation/tgstation/blob/9983b883f7c0b52697f0c9f3fe5b483cba25d4ff/code/modules/reagents/chemistry/machinery/chem_heater.dm#L449
Even though the debug chem heater has a capacity over 1000

https://github.com/tgstation/tgstation/blob/9983b883f7c0b52697f0c9f3fe5b483cba25d4ff/code/modules/reagents/chemistry/machinery/chem_heater.dm#L495-L497
Now the absorbing mechanism is based on the `maximum_volume` of the
holder and not on a fixed constant
4. Fixed the chem heater showing the wrong overlay when its panel is
opened and then closed while a beaker is inside
It would use the same icon for both open & closed state

https://github.com/tgstation/tgstation/blob/9983b883f7c0b52697f0c9f3fe5b483cba25d4ff/code/modules/reagents/chemistry/machinery/chem_heater.dm#L164
But it should actually use this state

https://github.com/tgstation/tgstation/blob/9983b883f7c0b52697f0c9f3fe5b483cba25d4ff/code/modules/reagents/chemistry/machinery/chem_heater.dm#L59
That's fixed now too.
5. Removed unused procs, variables and a whole lot of tracking which
UI's are opened/closed. All this is not necessary because the `SStgui`
subsystem tracks them for you so it was a lot of code bloat. autodocs
vars, procs and the regular stuff
6. Added more examines & tooltips for chem heater

## Changelog
:cl:
fix: debug chem heater now withdraws more than 100 units from its
inserted beaker.
fix: chem heater has correct overlay when its panel is closed with a
screwdriver while a beaker is inside.
qol: added more examines and tooltips for the chem heater.
code: converted chem heater UI to typescript. removed unused procs,
vars, ui tracking code. Added auto doc for everything.
refactor: removed chem heater tutorial help button & its related reward,
chem heater code has been optimized as a whole.
/:cl: