[MIRROR] [MDB IGNORE] Adds a brown variant of the Detective's Noir Suit (#9198)
Original PR: https://github.com/tgstation/tgstation/pull/80451
-----
## About The Pull Request
Adds a simple recoloured variant of the detective's nor suit into the
same brown as the detective's trenchcoat. So credit to PositiveEntropy
for the base noir suit sprite and the colour palette.

![image](https://github.com/tgstation/tgstation/assets/83892995/fb5b9022-1942-496e-8c69-a88c2fc1bae9)

![image](https://github.com/tgstation/tgstation/assets/83892995/63cb4f01-a43a-4c7a-a1c7-0bfb9510336a)


## Why It's Good For The Game
It never made sense to me why there were brown and black versions of the
detective suit and trench coats, but only a black suit jacket. To keep
the duality of the detective's colour scheme consistent, a brown version
has been added to compliment. Also it looks nice and allows further
expression for detectives.

## Changelog

:cl:
add: Adds a brown suit jacket to the detective's vendor
image: Adds a new sprite for the a brown detective suit jacket
:cl:

---------

Co-authored-by: ReinaCoder <83892995+ReinaCoder@users.noreply.github.com>