[MIRROR] [MDB IGNORE] A "postal workers strike" negative station trait. (can be positive) (#9209)
Original PR: https://github.com/tgstation/tgstation/pull/80213
-----
## About The Pull Request
Added a negative station trait, which blocks the mail like it's sunday
or a official holiday. However! If it's actually sunday or a official
holiday, when the mail would be normally blocked already, instead it'll
unblock it. It was done this way because it's easier (station traits are
loaded before the events and economy subsystems are initialized) and
cooler.

## Why It's Good For The Game
Granted there are quite a few station traits around cargo already,
features interacting with other features is honestly a good thing.

## Changelog

:cl:
add: Added a "postal workers strike" negative station trait. In the case
of holidays and sunday though, it'll be a "postal system overtime"
instead.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>