[MIRROR] [MDB IGNORE] Fixes AI behavior with breaking cuffs and resisting aggressive grabs (#9226)
Original PR: https://github.com/tgstation/tgstation/pull/80328
-----

## About The Pull Request

Makes it so that ai's recognize they're in a do_after after resisting,
preventing them from processing and interrupting the do_after while
they're in a condition that necessitates it. If it gets interrupted then
they'll process as normal, or if they finish they'll be free.
## Why It's Good For The Game

Monkeys can now resist things that they should've been, such as
aggressive grabs or cuffs, instead of having their ai completely freeze
when they're cuffed.
## Changelog
:cl:
fix: ai can now tell if it is in a do_after for resisting and will not
interrupt it. monkeys also now don't freeze up when aggressively grabbed
and will resist out of those and cuffs.
/:cl:

---------

Co-authored-by: Diamond_ <novadiamonddigger@gmail.com>