[MIRROR] [MDB IGNORE] Makes searchbars more responsive (#9270)
Original PR: https://github.com/tgstation/tgstation/pull/80688
-----

## About The Pull Request
Identified in Issue #80679: The current implementation of the search bar
component incorporates a timeout mechanism that's conflicting with the
controlled input behavior. This is leading to a race condition where
user input is inconsistently modified due to the timeout. To address
this, I've eliminated the timer and applied a uniform debounce of 200ms
to all onInput events. This approach aims to enhance input
responsiveness and prevent the erratic behavior previously observed.

Basic code maintenance: The search bar component itself is in
fabrication but used elsewhere, so I remade it as a functional component
and placed it in /common instead.
 
<details>
<summary>video</summary>


![mMiNInzCJp](https://github.com/tgstation/tgstation/assets/42397676/8567c1e6-21cf-4841-8080-78c35f47eb7b)

</details>
## Why It's Good For The Game
Protolathes are bearable again
Fixes #80679
## Changelog
:cl:
fix: Protolathe input should feel more responsive.
/:cl:

---------

Co-authored-by: Jeremiah <42397676+jlsnow301@users.noreply.github.com>