[MIRROR] [MDB IGNORE] [no gbp] adds options for bots to bypass adding unreachable items to their blacklist (#9313)
Original PR: https://github.com/tgstation/tgstation/pull/80708
-----

## About The Pull Request
bots will automatically add things it cant reach to a temporary
blacklist and ignore them until the blacklist is cleared. this however
can cause issues if u move out of the way of a stationary medbot while
its healing u, so it deems u unreachable and not try to heal u until the
blacklist is cleared

## Why It's Good For The Game
bots will automatically add things it cant reach to a temporary
blacklist and ignore them until the blacklist is cleared. this however
can cause issues if u move out of the way of a stationary medbot while
its healing u, so it deems u unreachable and not try to heal u until the
blacklist is cleared

## Changelog
:cl:
fix: stationary medbots will no longer ignore u for a while if u move
out of their way while healing
/:cl:

---------

Co-authored-by: Ben10Omintrix <138636438+Ben10Omintrix@users.noreply.github.com>