Fixes pod launcher BSOD [no gbp] (#80857)
## About The Pull Request
This fixes an issue caused by a class component and a pseudo hook called
useState.
The screen works now.

It's also terribly overdue for a refactor, which I'm working on after
looking at it. As a general rule I think tgui interfaces should keep it
under 500 lines. This one is 1203. I want to fix it before I do anything
else to it for file diff reasons.
## Why It's Good For The Game
Fixes #80850
## Changelog
:cl:
admin: Fixed the bluescreen in the centcom pod launcher.
/:cl: