[MIRROR] [MDB IGNORE] Stop misfiring from re-iterating and looping through every munition in the mag like that doesn't make much sense. (#9369)
Original PR: https://github.com/tgstation/tgstation/pull/80796
-----
## About The Pull Request
So, the misfiring proc (`blow_up`) can trigger even when the shot is
aimed at oneself, which is what the proc does. It also tries to loop
through every loaded projetile in the magazine when you only need the
chambered round, which is the one we're going to fire anyway. It's...
just plain dumb and can freeze your game for a fair bit.

## Why It's Good For The Game

![immagine](https://github.com/tgstation/tgstation/assets/42542238/f449e6ac-6fc5-4677-b409-c0e67ad3c137)

## Changelog

:cl:
fix: Fixed misfiring for firearms like tinkered detective revolvers.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>