[MIRROR] [MDB IGNORE] Removes infinite sources of plasma from icebox (#9542)
Original PR: https://github.com/tgstation/tgstation/pull/81091
-----

## About The Pull Request

The gibotonite that arcane had setup to spawn on icebox was uh, the
wrong kind, it was the cursed SNOWDIN gibonite, which is made to live in
a plasma environment. Rookie error really should have remembered that
ice/icemoon exists this is a GREAT codebase

Speaking of, I've noticed that gibonite is the only ore that spawns in
an actual mineral wall here (IE the blue rock walls) so you can really
trivially avoid it. that feels wrong

## Changelog
:cl:
fix: Icebox will no longer spawn a fuck ton of plasma after gibonite
blows. YW besties
/:cl:

Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>