[MIRROR] [MDB IGNORE] Shivs and bone daggers are no longer electrically conductive (#9576)
Original PR: https://github.com/tgstation/tgstation/pull/81145
-----
## About The Pull Request

Makes shivs and bone daggers not conduct electricity (i.e. you don't get
shocked from hitting an electrified grill with it).

In #80033 `CONDUCTS_ELECTRICITY` was moved to `obj_flags` instead of the
`flags_1` it was before and given a new constant for it. The author
generally replaced the `flags_1` with `obj_flags` but seems to have
forgotten to do this for the bone dagger and glass shiv `Initialize`
calls, while replacing the flag to the new constant though. Meaning that
they weren't non-conductive anymore.

## Why It's Good For The Game

These aren't meant to conduct electricity, this resolves that issue of
them doing that currently.

## Changelog

:cl:
fix: Bone daggers and shivs are now longer electrically conductive
/:cl:

---------

Co-authored-by: Inari-Whitebear <inari_whitebear@fastmail.jp>